# docs/api.yaml
openapi: 3.0.3
info:
  title: Kzquandary Mini Project - OpenAPI 3.0
  description: |-
    Mini Project for Alterra Academy with topic Rubbish Collection, user can create pickup request, topup balance, and chat. and collector can get transaction create transaction and chat
  contact:
    email: kaze@kzquandary.my.id
  version: 1.0.11
...
servers:
  - url: https://api.kzquandary.my.id/api/v1
...
paths:
...
  /transaction:
    post:
      tags:
        - transaction
      summary: Create Transaction
      description: Collector melakukan pickup
      operationId: transaction
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - pickup_id
              properties:
                pickup_id:
                  type: string
              example:
                pickup_request_id: 1234
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
              example:
                status: true
                message: Pickup Transaction Created
                data: 
                - id: 716553
                  pickup_request_id: 484900
                  pickup_time: "2024-05-09T22:21:00Z"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                example:
                  status: false
                  message: Pickup Transaction Failed
                  data: []
    get:
      tags:
        - transaction
      summary: Get Transaction
      description: Collector mendapatkan transaction
      operationId: getTransaction
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
              example:
                status: true
                message: Transaction Found
                data:
                  status: true
                  message: Pickup Transaction Found
                  data:
                    - id: 716553
                      user:
                        id: 239549
                        name: Kaze User
                        address: Cimahi
                      collector:
                        id: 958063
                        name: Kaze Collector
                      pickup_time: "2024-05-09T22:21:00Z"
                      tps_id: 1
                    - id: 716551
                      user:
                        id: 239549
                        name: Kaze User
                        address: Cimahi
                      collector:
                        id: 958063
                        name: Kaze Collector
                      pickup_time: "2024-05-09T22:21:00Z"
                      tps_id: 1
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                example:
                  status: false
                  message: Transaction not found
                  data: []
...
  /payments/verify/{order_id}:
    get:
      tags:
        - payment
      summary: Verify Payment
      description: Verify Payment
      operationId: verifyPayment
      security:
        - Bearer: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
          example: 106b83bd-8aa7-457c-b1f9-7fcdbe636a57
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                example:
                  status: true
                  message: Payment Verified
                  data: []
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                example:
                  status: false
                  message: Payment not found
                  data: []
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
                example:
                  status: false
                  message: Payment already verified
                  data: []
...